Rename query builder and metadata query
