+++ public/app/plugins/datasource/influxdb/components/VisualInfluxQLEditor/Editor.tags.test.tsx
@@ -3,12 +3,12 @@ import userEvent from '@testing-library/user-event';
 import React from 'react';
 
 import InfluxDatasource from '../../datasource';
-import * as mockedMeta from '../../influxQLMetadataQuery';
+import * as mockedMeta from '../../influxql_metadata_query';
 import { InfluxQuery } from '../../types';
 
 import { Editor } from './Editor';
 
-jest.mock('../../influxQLMetadataQuery', () => {
+jest.mock('../../influxql_metadata_query', () => {
   return {
     __esModule: true,
     getAllPolicies: jest.fn().mockReturnValueOnce(Promise.resolve(['default', 'autogen'])),
